+++ src/Normalization/ExtensionNormalization.php
@@ -41,6 +41,18 @@ class ExtensionNormalization implements NormalizationInterface
     ];
 
     /**
+     * These extensions are known and shouldn't be fully normalized.
+     *
+     * @var array<int,string>
+     */
+    private array $knownExtensions = [
+        // Files with uppercase .S extension include assembly code that needs to
+        // go through a pre-processor. The lowercase .s is assembly code that
+        // can be compiled into an object.
+        'S',
+    ];
+
+    /**
      * @var array<string,string>
      */
     private array $extensions = [
@@ -70,22 +82,37 @@ class ExtensionNormalization implements NormalizationInterface
             return $file;
         }
 
-        $newExtension = strtolower($extension);
+        $newExtension = $extension;
 
         // Trim special characters from extension.
         $newExtension = trim($newExtension, ' -');
+
+        // If the extension after trimming is known, leave it as it is.
+        if (in_array($newExtension, $this->knownExtensions, true)) {
+            $this->setFile($file, $extension, $newExtension);
+            return $file;
+        }
+
+        $newExtension = strtolower($newExtension);
 
         if (isset($this->extensions[$newExtension])) {
             $newExtension = $this->extensions[$newExtension];
         }
 
-        $basename = rtrim($file->getNewFilename(), $extension) . $newExtension;
+        $this->setFile($file, $extension, $newExtension);
+
+        return $file;
+    }
+
+    private function setFile(File $file, string $previousExtension, string $newExtension) {
+        $basename = rtrim($file->getNewFilename(), $previousExtension) . $newExtension;
 
         if ($basename !== $file->getNewFilename()) {
             $file->setNewFilename($basename);
-            $this->eventDispatcher->dispatch(new NormalizationEvent($file, 'file extension: ' . $extension . ' -> ' . $newExtension));
+            $this->eventDispatcher->dispatch(new NormalizationEvent(
+                $file,
+                'file extension: ' . $previousExtension . ' -> ' . $newExtension
+            ));
         }
-
-        return $file;
     }
 }